+++ .storybook/metrics/constant.js
@@ -1,0 +1,20 @@
+export const DEFAULT_ELEMENT = 0;
+export const CATEGORY = {
+    USERS: 'users',
+    PAGE: 'page',
+    EDITOR: 'editor',
+    TAB: 'tab',
+    SWITCHER: 'switcher',
+    PATH: 'path',
+    SEARCH_VALUE: 'search value',
+};
+
+export const ACTIONS = {
+    SEARCH_VALUE: 'Поиск через поисковую строку',
+    FOLDER: 'Сворачивание папки',
+};
+
+export const LINK = {
+    DIRECT: 'Прямая ссылка',
+    DEFAULT: 'Ссылка на главную страницу',
+};