+++ packages/custom-button/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 3.1.5
+
+### Patch Changes
+
+### [#588](https://github.com/core-ds/core-components/pull/588)
+
+-   Добавлен \_\_esModule в cjs экспорт
+
+-   Обновлены зависимости
+    -   button@8.0.0
+
 ## 3.1.4
 
 ### Patch Changes