+++ packages/date-input/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 4.2.1
+
+### Patch Changes
+
+### [#588](https://github.com/core-ds/core-components/pull/588)
+
+-   Добавлен \_\_esModule в cjs экспорт
+
+-   Обновлены зависимости
+    -   input@11.1.9
+
+## 4.2.0
+
+### Minor Changes
+
+### [#579](https://github.com/core-ds/core-components/pull/579)
+
+-   Добавлены новые правила форматирования для DateInput, DateRangeInput, DateTimeInput
+
 ## 4.1.8
 
 ### Patch Changes