+++ packages/icon-view/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.0.6
+
+### Patch Changes
+
+### [#588](https://github.com/core-ds/core-components/pull/588)
+
+-   Добавлен \_\_esModule в cjs экспорт
+
+### [#587](https://github.com/core-ds/core-components/pull/587)
+
+-   Добавлена переменная в css для фона в компоненте IconView
+
 ## 3.0.5
 
 ### Patch Changes