+++ packages/stepped-progress-bar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 2.0.12
+
+### Patch Changes
+
+### [#588](https://github.com/core-ds/core-components/pull/588)
+
+-   Добавлен \_\_esModule в cjs экспорт
+
+-   Обновлены зависимости
+    -   typography@3.2.1
+
 ## 2.0.11
 
 ### Patch Changes