+++ packages/underlay/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @alfalab/core-components-underlay
 
+## 1.0.1
+
+### Patch Changes
+
+### [#588](https://github.com/core-ds/core-components/pull/588)
+
+-   Добавлен \_\_esModule в cjs экспорт
+
+-   Обновлены зависимости
+    -   gap@1.1.2
+
 ## 1.0.0
 
 ### Major Changes